+++ app/files/changelog.md
@@ -1,3 +1,8 @@
+## Version Pf9.0.0 - February 2026
+- Ingested MalariaGEN Pf9 - now 40,074 QC-passed samples
+- Previous Pf-HaploAtlas Pf8 version has been made available [here](https://pf8-haploatlas.streamlit.app/) (it may require waking up the app!)
+
+
 ## Version Pf8.0.0 - November 2024
 - Ingested MalariaGEN Pf8 - now 24,409 QC-passed samples
 - Updated backend with new GFF - now 4,952 core genes